@@ -1,5 +1,5 @@
 plugins {
-    id 'org.kordamp.gradle.jacoco' version '0.43.0'
+    id 'org.kordamp.gradle.jacoco' version '+'
 }
 
 allprojects {
@@ -16,7 +16,7 @@ subprojects {
 	
 	group = 'oopp'
 	version = '0.0.1-SNAPSHOT'
-	sourceCompatibility = '11'
+	sourceCompatibility = '17'
 	checkstyle {
 		toolVersion "9.2.1"
 
