@@ -69,6 +69,7 @@ public class BoardList {
 
     public void removeCard(Card card) {
         this.cards.remove(card);
+        //card.boardList = null;
     }
 
     // other helper methods
@@ -79,13 +80,9 @@ public class BoardList {
 
     @Override
     public boolean equals(Object obj) {
-        if (obj instanceof BoardList) {
-            BoardList bl = (BoardList) obj;
+        if (obj instanceof BoardList bl) {
 
-            if (bl.cards.equals(this.cards) && bl.name.equals(this.name)) {
-                return true;
-            }
-            return false;
+            return bl.cards.equals(this.cards) && bl.name.equals(this.name);
         }
         return false;
     }
