@@ -14,7 +14,6 @@ import jakarta.ws.rs.WebApplicationException;
 import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.Node;
@@ -34,17 +33,14 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
+import java.util.Objects;
 import java.util.ResourceBundle;
-
-
-@SuppressWarnings("checkstyle:Indentation")
 
 public class SingleBoardCtrl implements Initializable {
     final ListCtrl listCtrl = new ListCtrl(this);
     final CardCtrl cardCtrl = new CardCtrl(this);
     final TagCtrl tagCtrl = new TagCtrl(this);
-    private BoardCtrl boardCtrl;
+    private final BoardCtrl boardCtrl;
     final SecurityCtrl securityCtrl = new SecurityCtrl(this);
     ServerUtils server;
     final BoardOverviewCtrl boardOverviewCtrl;
@@ -83,8 +79,6 @@ public class SingleBoardCtrl implements Initializable {
     Button copyInvite;
     boolean isUnlocked;
 
-    private LocalUtils localUtils;
-
     @Inject
     public SingleBoardCtrl(
             ServerUtils server,
@@ -96,14 +90,13 @@ public class SingleBoardCtrl implements Initializable {
         this.mainCtrl = mainCtrl;
         this.server = server;
         this.isUnlocked = isUnlocked;
-        this.localUtils = localUtils;
         this.boardCtrl = new BoardCtrl(this, localUtils);
     }
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-        ImageView imageView = new ImageView(getClass()
-                .getResource("../images/settings_icon.png")
+        ImageView imageView = new ImageView(Objects.requireNonNull(getClass()
+                        .getResource("../images/settings_icon.png"))
                 .toExternalForm());
         imageView.setFitWidth(settingsBtn.getPrefWidth());
         imageView.setFitHeight(settingsBtn.getPrefHeight());
@@ -146,7 +139,7 @@ public class SingleBoardCtrl implements Initializable {
         board_name.focusedProperty().addListener((obs, oldVal, newVal) -> {
             if (!newVal) {
                 try {
-                    boardCtrl.requestBoardName(board_name, BoardID);
+                    boardCtrl.requestBoardName(board_name);
                     refresh();
                 } catch (Exception e) {
                     refresh();
@@ -167,12 +160,10 @@ public class SingleBoardCtrl implements Initializable {
                 " -fx-background: transparent; -fx-border-color: transparent;");
         refresh();
         System.out.println(server);
-        server.checkForUpdatesToRefresh("/topic/lists", BoardList.class, boardList -> {
-            Platform.runLater(this::refresh);
-        });
-        server.checkForUpdatesToRefresh("/topic/boards", Board.class, board -> {
-            Platform.runLater(this::refresh);
-        });
+        server.checkForUpdatesToRefresh(
+                "/topic/lists", BoardList.class, boardList -> Platform.runLater(this::refresh));
+        server.checkForUpdatesToRefresh(
+                "/topic/boards", Board.class, board -> Platform.runLater(this::refresh));
     }
 
 
@@ -185,56 +176,13 @@ public class SingleBoardCtrl implements Initializable {
         return securityCtrl.checkReadOnlyMode(isUnlocked);
     }
 
-    void createNewList() {
-        listCtrl.createNewList();
-    }
-
-    public Node wrapList(BoardList boardList, ObservableList<Node> board_lists) throws IOException {
-        return listCtrl.wrapList(boardList, board_lists);
-    }
-
-    public void requestNameChange(TextField title, Node list) throws Exception {
-        listCtrl.requestNameChange(title, list);
-    }
-
-    void setListTitle(BoardList boardList, Node list) {
-        listCtrl.setListTitle(boardList, list);
-    }
-
-    void setDeleteBoardList(BoardList boardList, ObservableList<Node> board_lists,
-                            Node list) {
-        listCtrl.setDeleteBoardList(boardList, board_lists, list);
-    }
-
     public void placeCard(VBox parent, Card card) {
         cardCtrl.placeCard(parent, card);
-    }
-
-//    public void setCardDetail(Node cardNode, VBox parent) {
-//        cardCtrl.setCardDetail(cardNode, parent);
-//    }
-
-    void setDone(long listId, Card current, ActionEvent event) {
-        cardCtrl.setDone(listId, current, event);
     }
 
     public void addNewCard(VBox parent) {
         cardCtrl.addNewCard(parent);
     }
-
-
-    public Optional<String> showTitleDialog() {
-        return cardCtrl.showTitleDialog();
-    }
-
-    void setDragAndDrop(VBox parent, Node cardNode) {
-        cardCtrl.setDragAndDrop(parent, cardNode);
-    }
-
-    void addCardAtIndex(long sourceListId, int dropIndex, Card draggedCard) {
-        cardCtrl.addCardAtIndex(sourceListId, dropIndex, draggedCard);
-    }
-
 
     static void alertError(WebApplicationException e) {
         Alert alert = new Alert(Alert.AlertType.ERROR);
@@ -243,12 +191,7 @@ public class SingleBoardCtrl implements Initializable {
         alert.showAndWait();
     }
 
-
-    public long getListIndex(Long boardId, Long listId) {
-        return listCtrl.getListIndex(boardId, listId);
-    }
-
-    public void updateCardFromList(Long boardId, Long listId, Card current) {
+    public void updateCardFromList(Long listId, Card current) {
         try {
             server.updateCardFromList(listId, current);
         } catch (WebApplicationException e) {
@@ -256,7 +199,7 @@ public class SingleBoardCtrl implements Initializable {
         }
     }
 
-    public void saveCardToList(Long boardId, Long listId, Card current) {
+    public void saveCardToList(Long listId, Card current) {
         try {
             server.addCardToList(listId, current);
         } catch (WebApplicationException e) {
@@ -264,29 +207,18 @@ public class SingleBoardCtrl implements Initializable {
         }
     }
 
-    public void deleteCardFromList(Long boardId, Long listIdIndex, Card current) {
+    public void deleteCardFromList(Long listIdIndex
+            , Card current, Boolean permanent) {
         try {
-            server.deleteCardFromList(listIdIndex, current);
+            server.deleteCardFromList(listIdIndex, current, permanent);
         } catch (WebApplicationException e) {
             alertError(e);
         }
     }
 
-    public void setDelete(ActionEvent event, Node hbox, Card current, long listId) {
-        cardCtrl.setDelete(event, hbox, current, listId);
-    }
-
-    public void setCancel(ActionEvent event, Node hboxCard) {
-        cardCtrl.setCancel(event, hboxCard);
-    }
-
-    public void openBoardSettings() throws IOException {
-        boardCtrl.openBoardSettings(BoardID, new ConnectHomeCtrl(server, mainCtrl));
-    }
-
     public void setBoard(Board board) {
         this.current_board = board;
-        this.BoardID = board.getId();
+        BoardID = board.getId();
     }
 
     public String pullLists(Long id) {
@@ -334,8 +266,8 @@ public class SingleBoardCtrl implements Initializable {
         tagCtrl.enterTagName();
     }
 
-    public void addCustomTag(VBox root, Card current) {
-        tagCtrl.loadTagCard(root, current);
+    public void addCustomTag(VBox root) {
+        tagCtrl.loadTagCard(root);
     }
 
     public void openBoardTags(VBox tags, Card current) {
@@ -363,12 +295,7 @@ public class SingleBoardCtrl implements Initializable {
         return nodeCardMap;
     }
 
-    public Board getCurrent_board() {
-        return current_board;
-    }
-
     public HBox getHbox_lists() {
         return hbox_lists;
     }
 }
-
